feat: Adiciona suporte para CG080 e atualiza repositórios
Adicionada referência ao namespace `CSCore.Domain.Interfaces.CG.CG08X.CG080` em vários arquivos para incluir a interface `ICG080Repository`.
Atualizado `IUnitOfWorkV2` e sua implementação para suportar `ICG080Repository`.
Criado o serviço `ICG080Service` e sua implementação `CG080ServiceImpl` para operações CRUD na entidade `CG080`.
Adicionado o controlador `CG080Controller` para expor endpoints HTTP.
Alterada a interface `ICG080Repository` para herdar de `IRepositorioBase` e ajustado o tipo de `InCG080ID` para `long`.
Atualização no subprojeto `CSLB900_Cores`.

+++ CSCore.Ifs.CG/Repository/CG08X/CG080/CG080RepositoryImpl.cs
@@ -16,7 +16,7 @@ namespace CSCore.Ifs.CG.Repository.CG08X.CG080
             _appDbContext = appDbContext;
         }
 
-        public async Task<Osusr8dwCsicpCg080?> GetByIdAsync(int InTenantID, int InCG080ID)
+        public async Task<Osusr8dwCsicpCg080?> GetByIdAsync(int InTenantID, long InCG080ID)
         {
             return await _appDbContext.Osusr8dwCsicpCg080s
                 .AsNoTracking()
